+++ Libplanet.Tests/Blocks/BlockHeaderTest.cs
@@ -27,7 +27,9 @@ namespace Libplanet.Tests.Blocks
                 txHash: ImmutableArray<byte>.Empty,
                 hash: TestUtils.GetRandomBytes(32).ToImmutableArray(),
                 miner: ImmutableArray<byte>.Empty,
-                timestamp: future
+                timestamp: future,
+                preEvaluationHash: TestUtils.GetRandomBytes(32).ToImmutableArray(),
+                evaluationDigest: ImmutableArray<byte>.Empty
             );
 
             Assert.Throws<InvalidBlockTimestampException>(
@@ -52,7 +54,9 @@ namespace Libplanet.Tests.Blocks
                 timestamp: _fx.Next.Timestamp.ToString(
                     BlockHeader.TimestampFormat,
                     CultureInfo.InvariantCulture
-                )
+                ),
+                preEvaluationHash: TestUtils.GetRandomBytes(32).ToImmutableArray(),
+                evaluationDigest: ImmutableArray<byte>.Empty
             );
 
             Assert.Throws<InvalidBlockNonceException>(() =>
@@ -74,7 +78,9 @@ namespace Libplanet.Tests.Blocks
                 timestamp: _fx.Next.Timestamp.ToString(
                     BlockHeader.TimestampFormat,
                     CultureInfo.InvariantCulture
-                )
+                ),
+                preEvaluationHash: TestUtils.GetRandomBytes(32).ToImmutableArray(),
+                evaluationDigest: ImmutableArray<byte>.Empty
             );
 
             Assert.Throws<InvalidBlockIndexException>(() =>
@@ -94,7 +100,9 @@ namespace Libplanet.Tests.Blocks
                 txHash: ImmutableArray<byte>.Empty,
                 hash: TestUtils.GetRandomBytes(32).ToImmutableArray(),
                 miner: ImmutableArray<byte>.Empty,
-                timestamp: now.ToString(BlockHeader.TimestampFormat, CultureInfo.InvariantCulture)
+                timestamp: now.ToString(BlockHeader.TimestampFormat, CultureInfo.InvariantCulture),
+                preEvaluationHash: TestUtils.GetRandomBytes(32).ToImmutableArray(),
+                evaluationDigest: ImmutableArray<byte>.Empty
             );
 
             Assert.Throws<InvalidBlockDifficultyException>(() =>
@@ -109,7 +117,9 @@ namespace Libplanet.Tests.Blocks
                 txHash: ImmutableArray<byte>.Empty,
                 hash: TestUtils.GetRandomBytes(32).ToImmutableArray(),
                 miner: ImmutableArray<byte>.Empty,
-                timestamp: now.ToString(BlockHeader.TimestampFormat, CultureInfo.InvariantCulture)
+                timestamp: now.ToString(BlockHeader.TimestampFormat, CultureInfo.InvariantCulture),
+                preEvaluationHash: TestUtils.GetRandomBytes(32).ToImmutableArray(),
+                evaluationDigest: ImmutableArray<byte>.Empty
             );
 
             Assert.Throws<InvalidBlockDifficultyException>(() =>
@@ -124,7 +134,9 @@ namespace Libplanet.Tests.Blocks
                 txHash: ImmutableArray<byte>.Empty,
                 hash: TestUtils.GetRandomBytes(32).ToImmutableArray(),
                 miner: ImmutableArray<byte>.Empty,
-                timestamp: now.ToString(BlockHeader.TimestampFormat, CultureInfo.InvariantCulture)
+                timestamp: now.ToString(BlockHeader.TimestampFormat, CultureInfo.InvariantCulture),
+                preEvaluationHash: TestUtils.GetRandomBytes(32).ToImmutableArray(),
+                evaluationDigest: ImmutableArray<byte>.Empty
             );
 
             Assert.Throws<InvalidBlockTotalDifficultyException>(() =>
@@ -144,7 +156,9 @@ namespace Libplanet.Tests.Blocks
                 txHash: ImmutableArray<byte>.Empty,
                 hash: TestUtils.GetRandomBytes(32).ToImmutableArray(),
                 miner: ImmutableArray<byte>.Empty,
-                timestamp: now.ToString(BlockHeader.TimestampFormat, CultureInfo.InvariantCulture)
+                timestamp: now.ToString(BlockHeader.TimestampFormat, CultureInfo.InvariantCulture),
+                preEvaluationHash: TestUtils.GetRandomBytes(32).ToImmutableArray(),
+                evaluationDigest: ImmutableArray<byte>.Empty
             );
 
             Assert.Throws<InvalidBlockPreviousHashException>(() =>
@@ -159,7 +173,9 @@ namespace Libplanet.Tests.Blocks
                 txHash: ImmutableArray<byte>.Empty,
                 hash: TestUtils.GetRandomBytes(32).ToImmutableArray(),
                 miner: ImmutableArray<byte>.Empty,
-                timestamp: now.ToString(BlockHeader.TimestampFormat, CultureInfo.InvariantCulture)
+                timestamp: now.ToString(BlockHeader.TimestampFormat, CultureInfo.InvariantCulture),
+                preEvaluationHash: TestUtils.GetRandomBytes(32).ToImmutableArray(),
+                evaluationDigest: ImmutableArray<byte>.Empty
             );
 
             Assert.Throws<InvalidBlockPreviousHashException>(() =>