+++ Libplanet/HashDigest.cs
@@ -57,12 +57,27 @@ namespace Libplanet
         /// the same to the <see cref="Size"/> the hash algorithm
         /// (i.e., <typeparamref name="T"/> requires.</exception>
         public HashDigest(byte[] hashDigest)
+            : this((hashDigest ?? throw new ArgumentNullException(nameof(hashDigest)))
+                .ToImmutableArray())
         {
-            if (hashDigest == null)
-            {
-                throw new ArgumentNullException(nameof(hashDigest));
-            }
+        }
 
+        /// <summary>
+         /// Converts a <see cref="ImmutableArray"/> of <see cref="byte"/> array into a
+         /// <see cref="HashDigest{T}"/>.
+         /// </summary>
+         /// <param name="hashDigest">A <see cref="byte"/> array that encodes
+         /// a <see cref="HashDigest{T}"/>.  It must not be <c>null</c>,
+         /// and its <see cref="Array.Length"/> must be the same to
+         /// <see cref="Size"/>.</param>
+         /// <exception cref="ArgumentNullException">Thrown when the given
+         /// <paramref name="hashDigest"/> is <c>null</c>.</exception>
+         /// <exception cref="ArgumentOutOfRangeException">Thrown when the given
+         /// <paramref name="hashDigest"/>'s <see cref="ImmutableArray{T}.Length"/> is not
+         /// the same to the <see cref="Size"/> the hash algorithm
+         /// (i.e., <typeparamref name="T"/> requires.</exception>
+        public HashDigest(ImmutableArray<byte> hashDigest)
+        {
             if (hashDigest.Length != Size)
             {
                 string message =
@@ -74,7 +89,7 @@ namespace Libplanet
                 );
             }
 
-            _byteArray = hashDigest.ToImmutableArray();
+            _byteArray = hashDigest;
         }
 
         /// <summary>